+++ granule_ingester/tests/reading_processors/test_GridReadingProcessor.py
@@ -14,6 +14,7 @@
 # limitations under the License.
 
 import unittest
+import warnings
 from os import path
 
 import numpy as np
@@ -422,13 +423,15 @@ class TestCalendars(unittest.TestCase):
         tile = nexusproto.NexusTile()
 
         with xr.open_dataset(granule_path, decode_cf=True) as ds:
-            reading_processor = GridReadingProcessor(
-                [data_var],
-                lat_var,
-                lon_var,
-                time=time_var
-            )
-            tile = reading_processor._generate_tile(ds, dimensions_to_slices, tile)
+            with warnings.catch_warnings():
+                warnings.simplefilter('ignore', RuntimeWarning)
+                reading_processor = GridReadingProcessor(
+                    [data_var],
+                    lat_var,
+                    lon_var,
+                    time=time_var
+                )
+                tile = reading_processor._generate_tile(ds, dimensions_to_slices, tile)
             assert tile.tile.grid_tile.time
 
     def test_julian_calendar_tile(self):